frontend/QuizMatrix/src/app/forgot-password/forgot-password.component.spec.ts
import {
  ComponentFixture,
  TestBed,
  tick,
  fakeAsync,
  flush,
  flushMicrotasks,
} from '@angular/core/testing';
import { ForgotPasswordComponent } from './forgot-password.component';
import { AuthService } from '../services/auth/auth.service';
import { of, throwError } from 'rxjs';
import { ReactiveFormsModule } from '@angular/forms';
import { MatFormFieldModule } from '@angular/material/form-field';
import { MatInputModule } from '@angular/material/input';
import { Router } from '@angular/router';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';

describe('ForgotPasswordComponent', () => {
  let component: ForgotPasswordComponent;
  let fixture: ComponentFixture<ForgotPasswordComponent>;
  let authService: jasmine.SpyObj<AuthService>;
  let router: jasmine.SpyObj<Router>;

  beforeEach(() => {
    const authServiceSpy = jasmine.createSpyObj('AuthService', [
      'forgotPassword',
    ]);
    const routerSpy = jasmine.createSpyObj('Router', ['navigate']);

    TestBed.configureTestingModule({
      declarations: [ForgotPasswordComponent],
      imports: [
        ReactiveFormsModule,
        MatFormFieldModule,
        MatInputModule,
        BrowserAnimationsModule,
      ],
      providers: [
        { provide: AuthService, useValue: authServiceSpy },
        { provide: Router, useValue: routerSpy },
      ],
    });

    fixture = TestBed.createComponent(ForgotPasswordComponent);
    component = fixture.componentInstance;
    authService = TestBed.inject(AuthService) as jasmine.SpyObj<AuthService>;
    router = TestBed.inject(Router) as jasmine.SpyObj<Router>;

    component.ngOnInit();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should initialize form controls', () => {
    expect(component.formGroup.contains('email')).toBeTrue();
  });

  it('should handle error when forgotPassword method fails', fakeAsync(() => {
    const validEmail = 'test@example.com';
    authService.forgotPassword.and.returnValue(throwError('Mail failed'));

    component.formGroup.patchValue({ email: validEmail });
    component.submitForm_reset();
    tick();

    expect(authService.forgotPassword).toHaveBeenCalledWith(validEmail);
    expect(component.isSent).toBeFalse();

    flushMicrotasks();
    tick();

    expect(component.timer).toBeUndefined();
  }));
});
